Fix icon.erb ignoring for erb-formatter
We have a large 'when' block in the icon.erb file, and erb-formatter
makes it difficult to read. Although I tried to exclude the icon.erb
file, it appears to be wrong. At the moment, all files starting with the
letters 'i', 'c', 'o', 'n' are ignored, resulting in multiple files not
being formatted by the erb-formatter.

"views/inference/endpoint/index.erb" was ignored due to this issue. I
removed it from the new list becasue it has multiline string and
erb-formatter has bug for them.