Fix K8s Worker Node Pricing
We were calculating the worker node prices slightly off, because
the calculator was assuming that the nodes would get 40GB storage,
while in reality the default storage is calculated in the VM nexus
with `vm_size.storage_size_options.first`. This commit uses the
same logic in the UI option content generator.