Avoid 2N+1 query issue when deleting private subnet
This previously resulted in 2 queries for each vm attached to the
private subnet, one to get the strand, and the other check for
the semaphore.  It's possible to fix that with eager loading, but
it's faster to do the check in a single query, by excluding
the vms with the semaphore set or strand label set, and see if
there are any remaining.