Add port and database name to connection string
This is to be more explicit about the information in the connection string.
With the addition of pgBouncer, we need to specify the port anyway, which
leaves only the database name to be optional. For the sake of consistency,
I decided to include it in the connection string as well.