Emit useful metrics every 1000 respirate strands
The previous approach was to sample every 1000 strands and just emit
the time to acquire the lease.  This made it difficult to see problems
unless they were affecting all strands.  It made the P99 and similar
metrics worthless.

This adds a dedicated metrics thread to respirate, which continuously
pops the metrics queue.  After collecting results from 1000 strands,
it calculates and then emits the metrics.

The following metrics are emitted:

Strand delay metrics:

scan_delay :: Time between when the strand was scheduled, and when
              the scan query picked the strand up.
queue_delay :: Time between when the scan query picked up the strand,
               and when a worker thread started working on the strand.
lease_delay :: Time between when a worker thread started working on
               the strand, and when it acquired (or failed to acquire)
               the strand's lease.

Respirate internals metrics:

queue_size :: The size of the strand queue after the worker thread
              picked up the strand.  This is the backlog of strands
              waiting to be processed.
available_workers :: The number of idle worker threads that are
                     waiting to work on strands.  There should only be
                     an idle worker if the queue size is currently 0.

For the above metrics, we compute the average, median, P75, P85, P95,
P99, and maximum values.

In addition to the above metrics, there is one additional metric, the
lease_acquire_percentage. This is the percentage of strands where the
lease was successfully acquired.  For single respirate processes, or
normally running respirate partitioned processes (1 process per
partition), this should be 100.0. For multi-process, non-partitioned
respirate, this can be significantly lower, as multiple processes try
to process the same strand concurrently, and some fail to acquire the
lease.