Use cert_store in MinioClient
Previously we were relying on the ssl_ca_file api of Excon which
seems that it couldn't parse a chain of certs. In order to fix that
we will parse the certs and add them all to a cert_store and then
pass that to the client which fixes the issue.