Reduce unfairness of respirate processes from 50% to 0.000000023283%
With only 4 respirate processes, using a single leading hex character
was fine, as each process serviced 1/4 of the total (completely fair).
Earlier today, we moved to 6 respirate processes.  With a single
leading hex character, this will result in some processes servicing
3/16 of the total, and other processes servicing 2/16 of the total,
a 50% difference between the processes.

Switch from an array-based approach with a 1 hex character to a
calculation-based approach using the first 8 hex characters.  For 6 processes,
the uuid space will be split into the following 6 partitions:

"00000000-0000-0000-0000-000000000000"..."2aaaaaaa-0000-0000-0000-000000000000"
"2aaaaaaa-0000-0000-0000-000000000000"..."55555555-0000-0000-0000-000000000000"
"55555555-0000-0000-0000-000000000000"..."80000000-0000-0000-0000-000000000000"
"80000000-0000-0000-0000-000000000000"..."aaaaaaaa-0000-0000-0000-000000000000"
"aaaaaaaa-0000-0000-0000-000000000000"..."d5555555-0000-0000-0000-000000000000"
"d5555555-0000-0000-0000-000000000000".."ffffffff-ffff-ffff-ffff-ffffffffffff"

At most, there is a 1.0/(16**8) difference between processes, or
0.000000023283%.

In addition to being more fair, the code is also simpler.