Fix the `stripe_data` helper for the payment method
It returns a `Stripe::StripeObject`, which doesn’t have all the methods
from a hash.

It fails in production with the following exception:

    NoMethodError: undefined method 'slice' for #<Stripe::StripeObject:0x00007f9c27e9b4a0> (NoMethodError)

`to_h` returns a hash, but the keys are symbols, not strings. Since we
use string keys in the codebase, it's better to convert them before
slicing.