Remove lease IS NULL condition from dispatcher scan query
strand.lease was changed to NOT NULL a while back. Keeping
this in the scan query doesn't cause problems, because the
PostgreSQL query planner just removes the condition since
it is always true.  However, since it doesn't serve any
purpose, and misleads developers into thinking that lease
can be NULL, it should be removed.