Add Location to PostgresTimeline - Migration
This information is required since there can be timelines without a
leader. For AWS S3 integrated timelines, we also don't have a
MinioCluster record in the database, therefor, we should still be able
to somehow refer to the address of the S3 store. For that, we are adding
the reference of location to the timeline. This way, we can always refer
to the S3 account that holds the backups of this timeline.