Rhizome: Refactor StorageKeyEncryption.
We stored encrypted key and auth tag as two base64 encoded strings
separately for SPDK storage. In Ubiblk we store concatenation of them as
a single base64 encoded string (which was more convenient to use in Rust
libraries).

This change refactors StorageKeyEncryption and move base64 encode/decode
out of `wrap_key` & `unwrap_key`,  so they can also be reused for
Ubiblk.