Add migration for ipv4_address table
This table will contain all ipv4 addresses that can be assigned in
a specific range.  It will eventually allow for public IPv4 address
selection for a VM in a single query.

This takes the unusual approach of having the foreign key to the
address table reference address.cidr instead of address.id. The
reason for doing that is to allow for a constraint ensuring the
ip address is inside the cidr. That's significantly more challenging
to implement correctly if the foreign key constraint references
address.id.