Use ipv4_address table to speed up VmHost#ip4_random_vm_network
When selecting a random IPv4 address, take the universe of available
IPv4 addresses for the host (stored in the ipv4_address table), remove
the already assigned addresseses (in the assigned_vm_address table),
and return a random remaining address (if any).

To avoid potential problems, if the ipv4_address table has not yet
been populated for the VmHost's addresses, but there is an IPv4
subnet assigned to the host, use Clog.emit and fallback to the
previous implementation.  We can avoid the fallback if we do not
see production issues after a certain amount of time.

This should not be put into production until after the ipv4_address
table has been populated for existing Address instances.  However,
if it is, it should still work (as long as the ipv4_address migration
has been run), it will just be slower.