Remove MockStringWithExitstatus class
We already use the "Net::SSH::Connection::Session::StringWithExitstatus"
class returned by `session.exec!` in some tests, so there's no need for
an additional "MockStringWithExitstatus" class. It also raises the
following warning in CI.

```
/home/runner/work/ubicloud/ubicloud/spec/prog/kubernetes/upgrade_kubernetes_node_spec.rb:6: warning: method redefined; discarding old exitstatus
/home/runner/work/ubicloud/ubicloud/spec/prog/kubernetes/upgrade_kubernetes_node_spec.rb:6: warning: method redefined; discarding old exitstatus=
/home/runner/work/ubicloud/ubicloud/spec/prog/kubernetes/upgrade_kubernetes_node_spec.rb:8: warning: method redefined; discarding old initialize
/home/runner/work/ubicloud/ubicloud/spec/model/vm_host_spec.rb:9: warning: previous definition of initialize was here
```

We might add the "StringWithExitstatus" alias for it, but I couldn't get
it to run with freeze.