Add 15m timeout to CI
We have a few examples where CI tests get stuck and are canceled after 6
hours, which is the default timeout for GitHub Actions. Since our CI
runs in less than 10 minutes, it would be better to fail the job earlier
if it gets stuck instead of waiting for 6 hours. This also encourages us
to keep the CI fast and efficient.