Add a total delay metric
This shows the total delay between when the strand was scheduled
and when the lease was acquired. This gives a better overall view of
strand delay, and should be the primary metric for seeing how the
dispatcher is keeping up with strand volume.  The other delay
metrics are still useful in determining which parts are slow.