Consider current strand delay when looking for old strands
When there are a large number of strands awaiting processing, and all
respirate processes are overloaded and experiencing high levels of
strand delay, a fixed 5 seconds to consider strands old is suboptimal.

Calculate the current strand delay by taking the p95 of the most
recent total delay.  Multiply that by 20% to account for potential
rapid delay growth.  Then add 5 seconds.

This will increase the overall delay before old strands are picked up.
However, it should allow keeping a high lease acquire percentage even
if all respirate processes are experiencing strand processing delays.

Since each respirate process independently calculates the old strand
delay based on it's current strand delay, if some respirate processes
are not busy, and others are, the respirate processes that are not
busy will help process strands for the other processes that are busy.
Due to the random distribution of ubids (for non-st ubids), this isn't
likely, but it is possible, and will help in that case.