Make storage_size a required parameter
storage_size is one of the core parameters like size. Giving it a default feels
wrong. It is also not possible to use a single default value as valid storage_size
sizes vary by the instance size. So far we used smallest allowed storage size for
given the instance size as the default, but I think it is better to ask the user
to provide it explicitly.