Use NUL delimiter for xargs in github actions
Some `encoded_jit_config` values can interact badly with `xargs`'s
quoting rules, which are not of that of the shell, but of xargs's own,
similar design (see the first few paragraphs of the manual page)

In 29cbec28c, I packed the token in `xargs` to avoid it being emitted
in our log telemetry, as we don't save stdin, because we customarily
use it for secrets.

Because we're only packing a single value for a single invocation of
the subcommand, we can use the NUL byte delimitation mode of `xargs`
to evade its own interpretation of spaces and quotes, so long as
`encoded_jit_config` is not binary. That no NUL byte appears in stdin
is fine: the end-of-file condition when emitting the token is enough.