Split VM creation page into GPU/non-GPU if get_ff_gpu_vm feature flag is set
This can potentially make it easier to select a GPU instance.

With the feature flag set, instead of one button link to the
create page, there are two:

* The "Create Virtual Machine" link does not show any GPU options

* The "Create GPU Virtual Machine" only shows valid options for a
  GPU configuration

If users have a direct link to the old page, it works the same as
before, showing both GPU and non GPU options.

For this to work, you need to check for available GPUs earlier,
and use it to filter the families and locations.