Register deadline for runner in wait state
GitHub allows self-hosted runners to stay active for up to 5 days. So we
shouldn't have any runner in the "wait" state beyond that limit.

https://docs.github.com/en/actions/hosting-your-own-runners/managing-self-hosted-runners/usage-limits-for-self-hosted-runners

This change adds a safety check by registering a 5-day deadline for
runners in the wait state. With this, all labels in the runner prog
now have associated deadlines to help detect and recover from stuck
states.