Use AWS S3 SDK client for bucket operations
We used to use our MinIO client to interract with AWS S3, too. However,
that has a bunch of problems because we manually created that to talk to
our own minio clusters. Now, introducing multiple regions, we have to
either dive deeper to refactor the client or we could simply introduce
the aws sdk. I chose latter considering it's much cheaper for
engineering.