Base respirate metric delay times on lease time for expired leases
If a lease has expired, the delay generally shows as 120 seconds,
which is misleading. It should be fine to use the lease time instead
of the delay time in this case, because the respirate metrics are
supposed to check how long strands are waiting to be processed,
and expired leases have not been waiting to be processed, they were
processing and then failed.