Make respirate sleep time depend on current dispatcher state
In calculating the sleep time, consider:

* The current size of the queue
* How many strands are processed per second
* The percentage of available workers

Set an absolute minimum of 0.2 seconds, to not overload the database,
and an absolute maximum of 1 second, to not wait too long before
checking for strands.

Previously, the sleep time was always 1, so this can result in
5x the number of scan queries.  However, with the current queue
size ratio of 4, and typical strands per second of 1x-4x worker
threads, that isn't currently likely.  However, we will try decreasing
queue size ratio in the future, and during busy times, we may need
lower sleep time to prevent starvation (workers waiting on strand
queue, and strands waiting to be picked up by scan query, but scan
thread sleeping). Hence the reason to make sleep time depend on
dispatcher state.