Refactor AWS Firewall Update to check multiple permissions for the same range
This code had a bug. When there were multiple cidr definitions for the
same port ranges, it would only check the first cidr range. Now, we
perform the check for all of the cidr ranges and identify them
individually. We also refactor the mapping so that it is handled in a
single iteration.