Add dump_sequel_caches rake task
This converts the marshalled cache file to a text based file. The
reason to have this is it allows you to debug changes in cache
files.  If you are rebasing a commit and there is a conflict in
a cache file, you can run this before the commit, run this after
the commit, and then `diff -ru` the two directories, to see what
actually changed in the cache.