Don't use cached children in `donate`.
E2E tests failed because a `Test::Prog::Vm`'s strand went back from
`ping_google` to `verify_extra_disks`. `Test::Prog::Vm` is run in two
ways: Either by the scheduler or by a call to `donate` in
`Prog::Test::VmGroup.wait_verify_vms`.

We suspect that the issue we saw might be because of caching in
`donate`, so this is a try to fix that issue.