Have dispatcher check for shutdown after popping strand queue
Before, it did the check before popping the strand queue. However,
in a race condition, it could start processing a new strand after
shutdown is started.  I think that is only likely if the the queue
is full and the scan queue is blocked on pushing a job, but it
seems better to check right before processing.

If we are shutting down, we won't block on the strand queue, since
the shutdown process takes care of closing the queue.