Change respirate queue shutdown processing.
Push nil to metrics queue instead of closing it.  This still triggers
the metrics_thread to exit, but it avoids a ClosedQueueError in the
run_strand ensure block.

Change strand queue to use close instead of pushing nil. The strand
thread is a SizedQueue, not a regular Queue, and a full strand queue
could block the scan thread and prevent exiting during shutdown. We
can avoid this issue by closing the queue, and rescuing
ClosedQueueError when pushing to the strand queue in start_cohort.