Remove lease_clear_debug_snapshot for every strand checkin
This results in 3 extra queries per lease checkin:

* BEGIN
* SELECT
* COMMIT

Looking at the production logs, I do not believe the benefit of
the information justifies the cost.

In this code, num_updated can never be more than 1, because id
is a primary key. So if this fails, num_updated must be 0. There
are only two possibilities:

* The strand has been deleted (most likely)

* Something violated the lease, so the lease time changed, but the
  strand still exists.  In this case, I don't think we need to do
  the SELECT query before the UPDATE.  We can run it after the
  UPDATE, only if the UPDATE does not affect a row, and still get
  just as useful information.