Avoid recursion if VmHost#veth_pair_random_ip4_addr selects address in use
Instead of recursing and trying a new random address, keep trying the next
valid address until an unused address is found.  Carve out the top of the
range so we can be sure to always find a valid address before the end of
the range.

While I believe this is an uncommon operation (even for a fully loaded host
with 30 VMs, should only be a < 1/1000 chance of random address already
being used), I'd like to avoid recursive approaches based on randomness.

In addition to removing recursion, this also reduces object allocation
using constants.

This selects existing local_vetho_ip values using the association
dataset instead of the association array.  This is only called by the
allocator, and at the point it is called, the association is not cached,
so using the dataset and only selecting the needed column is faster and
uses less memory.