Decrement recycle if the server continues to apply WAL files
This can happen if the server stopped applying WAL files for a while, possibly
due to incompatible config changes. Then, when same config changes are applied
to the read replica, the server may continue applying WAL files again. In this
case, we should stop recycling the server and give it a chance to catch up.